renamed component for match requirement and update jwt when user modify jwt user will get denied from server and user will get logged out

+++ src/Hooks/useMyItems/useMyItemsProducts.js
@@ -4,7 +4,8 @@ import useUser from "../useFirebase/useUser";
 
 const useMyItemsProducts = () => {
     const [products, setProducts] = useState([]);
-    const { user } = useUser();
+    const [loading, setLoading] = useState(true);
+    const { user, handleLogout } = useUser();
 
     useEffect(() => {
         if (user?.email) {
@@ -19,31 +20,37 @@ const useMyItemsProducts = () => {
                 .then(data => {
                     if (Array.isArray(data)) {
                         setProducts(data);
+                        setLoading(false);
                     }
                     else {
                         toast.error(data.message);
+                        handleLogout();
                     }
                 })
         }
 
-    }, [user]);
+    }, [user, handleLogout]);
 
     const handleDelete = (id) => {
+        const confirm = window.confirm("Are you sure you want to delete this product?");
         const url = `https://sabbir-assignment-11.herokuapp.com/api/products/${id}`;
 
-        fetch(url, {
-            method: 'delete',
-            headers: {
-                'Content-Type': 'application/json'
-            }
-        })
-            .then(res => res.json())
-            .then(data => console.log(data))
-        toast.info('product deleted successfully')
-        setProducts(products.filter(product => product._id !== id));
+        if (confirm) {
+            fetch(url, {
+                method: 'delete',
+                headers: {
+                    'Content-Type': 'application/json'
+                }
+            })
+                .then(res => res.json())
+            toast.info('product deleted successfully')
+            setProducts(products.filter(product => product._id !== id));
+        } else {
+            toast.info('product not deleted')
+        }
     }
 
-    return { products, handleDelete };
+    return { products, loading, handleDelete };
 };
 
 export default useMyItemsProducts;
